+++ Source/Core/Visualization/Event/EventListener.h
@@ -26,36 +26,36 @@ namespace kvs
 class EventListener
 {
 public:
-    using InitializeEvent = std::function<void()>;
-    using PaintEvent = std::function<void()>;
-    using ResizeEvent = std::function<void(int,int)>;
-    using MousePressEvent = std::function<void(kvs::MouseEvent*)>;
-    using MouseMoveEvent = std::function<void(kvs::MouseEvent*)>;
-    using MouseReleaseEvent = std::function<void(kvs::MouseEvent*)>;
-    using MouseDoubleClickEvent = std::function<void(kvs::MouseEvent*)>;
-    using WheelEvent = std::function<void(kvs::WheelEvent*)>;
-    using KeyPressEvent = std::function<void(kvs::KeyEvent*)>;
-    using KeyRepeatEvent = std::function<void(kvs::KeyEvent*)>;
-    using KeyReleaseEvent = std::function<void(kvs::KeyEvent*)>;
-    using TimerEvent = std::function<void(kvs::TimeEvent*)>;
+    using InitializeEventFunc = std::function<void()>;
+    using PaintEventFunc = std::function<void()>;
+    using ResizeEventFunc = std::function<void(int,int)>;
+    using MousePressEventFunc = std::function<void(kvs::MouseEvent*)>;
+    using MouseMoveEventFunc = std::function<void(kvs::MouseEvent*)>;
+    using MouseReleaseEventFunc = std::function<void(kvs::MouseEvent*)>;
+    using MouseDoubleClickEventFunc = std::function<void(kvs::MouseEvent*)>;
+    using WheelEventFunc = std::function<void(kvs::WheelEvent*)>;
+    using KeyPressEventFunc = std::function<void(kvs::KeyEvent*)>;
+    using KeyRepeatEventFunc = std::function<void(kvs::KeyEvent*)>;
+    using KeyReleaseEventFunc = std::function<void(kvs::KeyEvent*)>;
+    using TimerEventFunc = std::function<void(kvs::TimeEvent*)>;
 
 private:
     int m_event_type; ///< event type
     std::string m_name; ///< name
     kvs::ScreenBase* m_screen; ///< pointer to the screen
     kvs::Scene* m_scene; ///< pointer to the scene
-    InitializeEvent m_initialize_event;
-    PaintEvent m_paint_event;
-    ResizeEvent m_resize_event;
-    MousePressEvent m_mouse_press_event;
-    MouseMoveEvent m_mouse_move_event;
-    MouseReleaseEvent m_mouse_release_event;
-    MouseDoubleClickEvent m_mouse_double_click_event;
-    WheelEvent m_wheel_event;
-    KeyPressEvent m_key_press_event;
-    KeyRepeatEvent m_key_repeat_event;
-    KeyReleaseEvent m_key_release_event;
-    TimerEvent m_timer_event;
+    InitializeEventFunc m_initialize_event;
+    PaintEventFunc m_paint_event;
+    ResizeEventFunc m_resize_event;
+    MousePressEventFunc m_mouse_press_event;
+    MouseMoveEventFunc m_mouse_move_event;
+    MouseReleaseEventFunc m_mouse_release_event;
+    MouseDoubleClickEventFunc m_mouse_double_click_event;
+    WheelEventFunc m_wheel_event;
+    KeyPressEventFunc m_key_press_event;
+    KeyRepeatEventFunc m_key_repeat_event;
+    KeyReleaseEventFunc m_key_release_event;
+    TimerEventFunc m_timer_event;
 
 public:
     EventListener( const int event_type = kvs::EventBase::AllEvents );
@@ -72,18 +72,18 @@ public:
     void setScene( kvs::Scene* scene ) { m_scene = scene; }
     void addEventType( const int event_type ) { m_event_type |= event_type; }
 
-    void initializeEvent( InitializeEvent e ) { m_initialize_event = e; }
-    void paintEvent( PaintEvent e ) { m_paint_event = e; }
-    void resizeEvent( ResizeEvent e ) { m_resize_event = e; }
-    void mousePressEvent( MousePressEvent e ) { m_mouse_press_event = e; }
-    void mouseMoveEvent( MouseMoveEvent e ) { m_mouse_move_event = e; }
-    void mouseReleaseEvent( MouseReleaseEvent e ) { m_mouse_release_event = e; }
-    void mouseDoubleClickEvent( MouseDoubleClickEvent e ) { m_mouse_double_click_event = e; }
-    void wheelEvent( WheelEvent e ) { m_wheel_event = e; }
-    void keyPressEvent( KeyPressEvent e ) { m_key_press_event = e; }
-    void keyRepeatEvent( KeyRepeatEvent e ) { m_key_repeat_event = e; }
-    void keyReleaseEvent( KeyReleaseEvent e ) { m_key_release_event = e; }
-    void timerEvent( TimerEvent e ) { m_timer_event = e; }
+    void initializeEvent( InitializeEventFunc func ) { m_initialize_event = func; }
+    void paintEvent( PaintEventFunc func ) { m_paint_event = func; }
+    void resizeEvent( ResizeEventFunc func ) { m_resize_event = func; }
+    void mousePressEvent( MousePressEventFunc func ) { m_mouse_press_event = func; }
+    void mouseMoveEvent( MouseMoveEventFunc func ) { m_mouse_move_event = func; }
+    void mouseReleaseEvent( MouseReleaseEventFunc func ) { m_mouse_release_event = func; }
+    void mouseDoubleClickEvent( MouseDoubleClickEventFunc func ) { m_mouse_double_click_event = func; }
+    void wheelEvent( WheelEventFunc func ) { m_wheel_event = func; }
+    void keyPressEvent( KeyPressEventFunc func ) { m_key_press_event = func; }
+    void keyRepeatEvent( KeyRepeatEventFunc func ) { m_key_repeat_event = func; }
+    void keyReleaseEvent( KeyReleaseEventFunc func ) { m_key_release_event = func; }
+    void timerEvent( TimerEventFunc func ) { m_timer_event = func; }
 
     virtual void onEvent( kvs::EventBase* event );
     virtual void initializeEvent() { if ( m_initialize_event ) m_initialize_event(); }