+++ Source/Core/Visualization/Event/MouseDoubleClickEventListener.h
@@ -21,10 +21,10 @@ class MouseDoubleClickEventListener : public kvs::EventListener
 {
 public:
     MouseDoubleClickEventListener(): kvs::EventListener( kvs::EventBase::MouseDoubleClickEvent ) {}
-    MouseDoubleClickEventListener( MouseDoubleClickEvent e ) { this->update( e ); }
+    MouseDoubleClickEventListener( MouseDoubleClickEventFunc func ) { this->update( func ); }
     virtual ~MouseDoubleClickEventListener() {}
 
-    void update( MouseDoubleClickEvent e ) { mouseDoubleClickEvent( e ); }
+    void update( MouseDoubleClickEventFunc func ) { mouseDoubleClickEvent( func ); }
     virtual void update( kvs::MouseEvent* event ) { mouseDoubleClickEvent( event ); }
 
 private: