+++ Source/Core/Visualization/Event/ScreenCaptureEvent.cpp
@@ -19,12 +19,29 @@ namespace kvs
 /*===========================================================================*/
 /**
  *  @brief  Constructs a new ScreenCaptureEvent class.
+ *  @param  key [in] key for capturing
  */
 /*===========================================================================*/
-ScreenCaptureEvent::ScreenCaptureEvent():
-    m_key( kvs::Key::s ),
+ScreenCaptureEvent::ScreenCaptureEvent( const int key ):
+    m_key( key ),
     m_filename( "" ),
-    m_basename( "screenshot" )
+    m_basename( "screenshot" ),
+    m_capture_func( nullptr )
+{
+}
+
+/*===========================================================================*/
+/**
+ *  @brief  Constructs a new ScreenCaptureEvent class.
+ *  @param  func [in] capture function
+ *  @param  key [in] key for capturing
+ */
+/*===========================================================================*/
+ScreenCaptureEvent::ScreenCaptureEvent( CaptureFunc func, const int key ):
+    m_key( key ),
+    m_filename( "" ),
+    m_basename( "screenshot" ),
+    m_capture_func( func )
 {
 }
 
@@ -40,16 +57,28 @@ void ScreenCaptureEvent::update( kvs::KeyEvent* event )
 
     if ( event->key() == m_key )
     {
-        std::string filename = m_filename;
-        if ( filename.empty() )
-        {
-            kvs::Date today;
-            kvs::Time now;
-            filename = m_basename + "_" + today.toString("") + now.toString("") + ".bmp";
-        }
-
-        scene()->camera()->snapshot().write( filename );
+        auto image = scene()->camera()->snapshot();
+        if ( m_capture_func ) { m_capture_func( image ); }
+        else { image.write( this->output_filename() ); }
     }
+}
+
+/*===========================================================================*/
+/**
+ *  @brief  Returns output image filename.
+ *  @return output image filename
+ */
+/*===========================================================================*/
+std::string ScreenCaptureEvent::output_filename() const
+{
+    std::string filename = m_filename;
+    if ( filename.empty() )
+    {
+        kvs::Date today;
+        kvs::Time now;
+        filename = m_basename + "_" + today.toString("") + now.toString("") + ".bmp";
+    }
+    return filename;
 }
 
 } // end of namespace kvs