+++ Source/Core/Visualization/Event/ScreenCaptureEvent.h
@@ -7,6 +7,8 @@
 #pragma once
 #include <kvs/KeyPressEventListener>
 #include <string>
+#include <functional>
+#include <kvs/ColorImage>
 
 
 namespace kvs
@@ -19,18 +21,27 @@ namespace kvs
 /*===========================================================================*/
 class ScreenCaptureEvent : public kvs::KeyPressEventListener
 {
+public:
+    using CaptureFunc = std::function<void(const kvs::ColorImage&)>;
+
 private:
     int m_key; ///< key for capturing screen image
     std::string m_filename; ///< filename of captured image
     std::string m_basename; ///< basename of captured image
+    CaptureFunc m_capture_func;
 
 public:
-    ScreenCaptureEvent();
+    ScreenCaptureEvent( const int key = kvs::Key::s );
+    ScreenCaptureEvent( CaptureFunc func, const int key = kvs::Key::s );
 
     void setKey( const int key ) { m_key = key; }
     void setFilename( const std::string& filename ) { m_filename = filename; }
     void setBasename( const std::string& basename ) { m_basename = basename; }
+    void update( CaptureFunc func ) { m_capture_func = func; }
     void update( kvs::KeyEvent* event );
+
+private:
+    std::string output_filename() const;
 };
 
 } // end of namespace kvs